Created tests: E2E tests (vue-app/tests/e2e/genre-management.spec.ts): Genre management page navigation, UI element verification, mock authentication Unit tests (vue-app/tests/unit/components/Genre.test.ts): Genre component rendering, form interactions, genre filtering Test config (vue-app/playwright.config.ts, vue-app/vitest.config.ts): Playwright and Vitest configurations Backend tests (tests/ folder): User service, genre service, and route testing with pytest. Updated config files.

+++ app/models/genre.py
@@ -1,9 +1,9 @@
-from sqlalchemy import Column, BigInteger, String, Text
+from sqlalchemy import Column, Integer, String, Text
 from app.db.session import Base
 
 class Genre(Base):
     __tablename__ = "genres"
 
-    id = Column(BigInteger, primary_key=True, index=True, autoincrement=True)
+    id = Column(Integer, primary_key=True, index=True, autoincrement=True)
     name = Column(String(256), unique=True, nullable=False)
     description = Column(Text, nullable=True)
